0.3.2 updates
Changed month format to short.
Added padding to text on post.
No min max height for thumbnail/post title image

+++ s3-webapp/src/js/utils.js
@@ -2,7 +2,7 @@ export const name = 'utils';
 
 function formatTimeStamp(str) {
     let time = '';
-    var options = {year: 'numeric', month: 'long', day: 'numeric', hour: '2-digit', 
+    var options = {year: 'numeric', month: 'short', day: 'numeric', hour: '2-digit', 
     minute:'2-digit', hour12: false};
     if (str) {
       time = new Date(str);
